% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AB}
\alias{AB}
\title{Example dataset containing some traits}
\format{
\subsection{\code{AB}}{

A data.frame with 102 columns and 96 rows:
\describe{
\item{Plot}{Factor w/ 8 levels "4","6","11",..: 2 4 6 8 2 4 6 8 2 ...}
\item{Year}{Factor w/ 12 levels "2010","2011",..: 1 1 1 1 2 2 2 2 3 3 ...}
\item{Treatment}{Factor w/ 2 levels "Mown_NPK",..: 1 1 1 1 1 1 1 1 1 ...}
\item{Bloc}{Factor w/ 2 levels "1","2": 1 1 2 2 1 1 2 2 1 }
\item{SP1}{num 0, 0.84, 0, 9.15, 0, ...}
\item{SP3}{num  0, 0.84, 0, 0, 0, ...}
...
}
}
}
\usage{
AB
}
\description{
Example dataset containing some traits
}
\keyword{internal}
