% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance_dataframe}
\alias{abundance_dataframe}
\title{Example dataset of abundance_dataframe}
\format{
\subsection{`abundance_dataframe}{

A data.frame with 2016 rows and 99 columns:
\describe{
\item{number}{int  0 0 0 0 0 0 0 0 0 0 ...}
\item{index1}{num  0 0.84 0 9.15 0 ...}
\item{index2}{num  0 0.84 0 0 0 ...}
\item{index3}{num  0 0 0 0 0 0 0 0 0 0 ...}
...
}
Example output of get_abundance_df
}
}
\usage{
abundance_dataframe
}
\description{
Example dataset of abundance_dataframe
}
\keyword{internal}
