% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{build_skr_ses}
\alias{build_skr_ses}
\title{skr ses genration/save/load}
\usage{
build_skr_ses(
  statistics_factor_name,
  significativity_threshold,
  stat_per_rand_file = NULL,
  skr_param = NULL,
  stat_skr_param_file = NULL,
  regenerate_ses_skr = FALSE,
  skew_uniform = FALSE
)
}
\description{
skr ses genration/save/load
}
\keyword{internal}
