% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_parameter_value}
\alias{check_parameter_value}
\title{check_parameter_value}
\usage{
check_parameter_value(value, checker, expected_description, fmt = NULL)
}
\description{
check_parameter_value
}
\keyword{internal}
