% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\docType{data}
\name{CONSTANTS}
\alias{CONSTANTS}
\title{The CONSTANTS constant}
\format{
An object of class \code{list} of length 6.
}
\usage{
CONSTANTS
}
\description{
Provides a set of constants to prevent typo and provide some defauts values
to functions in the TAD.
Among those constants are:
\itemize{
\item SKEW_UNIFORM_SLOPE_DISTANCE
\item SKEW_UNIFORM_INTERCEPT_DISTANCE
\item DEFAULT_SIGNIFICATIVITY_THRESHOLD
\item DEFAULT_LIN_MOD
\item DEFAULT_SLOPE_DISTANCE
\item DEFAULT_INTERCEPT_DISTANCE
}
}
\keyword{datasets}
