% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{filter_na_empty}
\alias{filter_na_empty}
\title{input filter}
\usage{
filter_na_empty(abundance_df, weights, weights_factor, trait_data)
}
\arguments{
\item{weights}{the dataframe of weights, one row correspond to a
series of observation}

\item{weights_factor}{the dataframe which contains the different
factor linked to the weights}

\item{trait_data}{a vector of the data linked to the different factor}
}
\description{
input filter
}
\keyword{internal}
