% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_abundances}
\alias{filtered_abundances}
\title{Example dataset of filtred results just after abundances generation}
\format{
An object of class \code{list} of length 4.
}
\usage{
filtered_abundances
}
\description{
Example dataset of filtred results just after abundances generation
}
\keyword{internal}
