% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{generate_stat_per_rand}
\alias{generate_stat_per_rand}
\title{stats per random generation}
\usage{
generate_stat_per_rand(
  weights,
  statistics_factor_name,
  weights_factor,
  weighted_moments,
  randomization_number,
  abundance_df,
  lin_mod
)
}
\arguments{
\item{weights}{the dataframe of weights, one row correspond to a
series of observation}

\item{statistics_factor_name}{vector of factor name for the
computation of statistics for each generated matrix}

\item{weights_factor}{the dataframe which contains the different
factor linked to the weights}

\item{randomization_number}{the number of random abundance matrix to
generate}

\item{lin_mod}{Indicates the type of linear model to use for
(SKR): choose "lm" or "mblm"}
}
\description{
stats per random generation
}
\keyword{internal}
