% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{get_abundance_df}
\alias{get_abundance_df}
\title{abundance generation}
\usage{
get_abundance_df(
  weights,
  weights_factor,
  randomization_number,
  abundance_file,
  regenerate_abundance_df,
  aggregation_factor_name,
  seed
)
}
\arguments{
\item{weights}{the dataframe of weights, one row correspond to a
series of observation}

\item{weights_factor}{the dataframe which contains the different
factor linked to the weights}

\item{randomization_number}{the number of random abundance matrix to
generate}

\item{abundance_file}{the path and name of the RDS file to
load/save the dataframe which
contains the observed data and the generated matrix}

\item{regenerate_abundance_df}{boolean to specify if the
abundance dataframe is computed again}

\item{aggregation_factor_name}{vector of factor name for the
generation of random matrix}

\item{seed}{the seed of the pseudo random number generator}
}
\description{
abundance generation
}
\keyword{internal}
