% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{get_stat_per_rand}
\alias{get_stat_per_rand}
\title{stats per random genration/save/load}
\usage{
get_stat_per_rand(
  weights,
  stat_per_rand_file,
  regenerate_stat_per_rand_df,
  statistics_factor_name,
  weights_factor,
  randomization_number,
  weighted_moments,
  abundance_df,
  lin_mod
)
}
\description{
stats per random genration/save/load
}
\keyword{internal}
