% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{get_weighted_mnts}
\alias{get_weighted_mnts}
\title{weighted moments generation}
\usage{
get_weighted_mnts(
  weights_factor,
  trait_data,
  weighted_moments_file,
  regenerate_weighted_moments_df,
  abundance_df,
  randomization_number,
  slope_distance,
  intercept_distance,
  factor_names = NULL
)
}
\arguments{
\item{weights_factor}{the dataframe which contains the different
factor linked to the weights}

\item{trait_data}{a vector of the data linked to the different factor}

\item{weighted_moments_file}{the path and name of the RDS file to
load/save the dataframe which
contains the calculated moments}

\item{regenerate_weighted_moments_df}{boolean to specify if
the weighted moments dataframe is computed again}

\item{randomization_number}{the number of random abundance matrix to
generate}

\item{slope_distance}{slope of the theoretical distribution
law (default: slope = 1 intercept = 1.86 skew-uniform distribution family)}

\item{intercept_distance}{intercept of the theoretical distribution
law (default: slope = 1 intercept = 1.86 skew-uniform distribution family)}
}
\description{
weighted moments generation
}
\keyword{internal}
