% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{load_abundance_dataframe}
\alias{load_abundance_dataframe}
\title{load_abundance_dataframe}
\usage{
load_abundance_dataframe(path, ...)
}
\arguments{
\item{path}{the path to the file to load}

\item{...}{a set of parameters provided to load_depending_on_format
may contain some operations to apply to format/cast CSV or TSV data
which are almost typeless by default}
}
\value{
an abundance dataframe, with the column \code{number} caster
into integers and rownames casted into integers.
}
\description{
load_abundance_dataframe
}
