% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{load_depending_on_format}
\alias{load_depending_on_format}
\title{load_depending_on_format}
\usage{
load_depending_on_format(path, env_attr = NULL, ...)
}
\arguments{
\item{path}{the path of the file to load}

\item{env_attr}{the name of the object to load}

\item{...}{only used in case of CSV loading - passed to load_tad_table
see load_tad_table$\code{...} for more informations}
}
\description{
load_depending_on_format
}
\keyword{internal}
