% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{load_stat_skr_param}
\alias{load_stat_skr_param}
\title{load_stat_skr_param}
\usage{
load_stat_skr_param(path, ...)
}
\arguments{
\item{path}{the path to the file to load}

\item{...}{a set of parameters provided to load_depending_on_format
may contain some operations to apply to format/cast CSV or TSV data
which are almost typeless by default}
}
\value{
a stats SKR parameters dataframe with distance_to_family_ses and
cv_distance_to_family_ses casted into doubles and with rownames casted
into integers.
}
\description{
load_stat_skr_param
}
