% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{load_statistics_per_obs}
\alias{load_statistics_per_obs}
\title{load_statistics_per_obs}
\usage{
load_statistics_per_obs(path, ...)
}
\arguments{
\item{path}{the path to the file to load}

\item{...}{a set of parameters provided to load_depending_on_format
may contain some operations to apply to format/cast CSV or TSV data
which are almost typeless by default}
}
\value{
a stats par observations dataframe with rownames casted
into integers.
}
\description{
load_statistics_per_obs
}
