% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{load_tad_table}
\alias{load_tad_table}
\title{load_tad_table}
\usage{
load_tad_table(
  path,
  sep,
  table_name = NULL,
  factor_names = NULL,
  character_names = NULL,
  ...
)
}
\arguments{
\item{path}{the path to the file to load}

\item{sep}{the separator used for this csv/tsv file}

\item{table_name}{the name of the object to load}

\item{factor_names}{provide the name of factor columns. A \code{as.factor}
function will be applied to the provided columns.}

\item{character_names}{provide the name of character columns. A
\code{as.character} function will be applied to the provided columns.}

\item{...}{a set of functions to apply to the result's attribute.
if \code{a} = as.numeric in \code{...}, then \code{result$a <- as.numeric(result$a)}
any other function than as.numeric can be used}
}
\description{
load_tad_table
}
\keyword{internal}
