% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{save_depending_on_format}
\alias{save_depending_on_format}
\title{save_depending_on_format}
\usage{
save_depending_on_format(file, var_name, compress = "bzip2")
}
\arguments{
\item{file}{the path of the file to save}

\item{var_name}{the name of the object to save}

\item{compress}{compression algo for rds/rda, defaults to bzip2}
}
\value{
NULL - called for side effects
}
\description{
Saves data in TSV, CSV, RDA or RDS, depending on the file's extension.
CSV and TSVs are saved the same way:
with double quotes (ascii character 34, or 0x22 " ),
EOL is a newline (ascii character 10, or 0x0A),
NAs are saved as "NA",
decimal point are latin point (ascii character 46, or 0x2E . ),
row names and colnames are added and the encoding is UTF-8.
The only thing that changes is the separator:
A comma (ascii character 44, or 0x2C , ) for CSV files and
a tabulation (ascii character 9, or 0x09) for TSV files.
}
\keyword{internal}
