% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{save_obs_df}
\alias{save_obs_df}
\title{observations genration/save/load}
\usage{
save_obs_df(
  weights_factor,
  stat_per_obs_file,
  regenerate_stat_per_obs_df,
  weighted_moments,
  randomization_number,
  significativity_threshold
)
}
\arguments{
\item{weights_factor}{the dataframe which contains the different
factor linked to the weights}

\item{stat_per_obs_file}{the path and name of the RDS file to
load/save the dataframe which
contains the statistics for each observed row regarding the random ones}

\item{regenerate_stat_per_obs_df}{boolean to specify if
the statistics per observation dataframe is computed again}

\item{randomization_number}{the number of random abundance matrix to
generate}

\item{significativity_threshold}{the significance threshold to
consider that the observed value is in the randomized value}
}
\description{
observations genration/save/load
}
\keyword{internal}
