% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{save_statistics_per_obs}
\alias{save_statistics_per_obs}
\title{save_statistics_per_obs}
\usage{
save_statistics_per_obs(path, object = NULL)
}
\arguments{
\item{path}{the path of the file to load}

\item{object}{the object to save}
}
\value{
NULL - called for side effects
}
\description{
This function provides a secured way to save statistics_per_obs
dataframe.
The more generic functionprovided by TAD \code{save_depending_on_format} expects
saves object using their name, but saves nothing if the provided name is not
correct, or mya even save an unwanted object.
This function provides a way to verify the object you want to save, and so,
it is more secured.
}
