% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skr_ses_dataframe}
\alias{skr_ses_dataframe}
\title{Example dataset of skr_ses}
\format{
\subsection{\code{skr_ses}}{

A data.frame - todo = describe the df content:
Example output of skr_ses
}
}
\usage{
skr_ses_dataframe
}
\description{
Example dataset of skr_ses
}
\keyword{internal}
