% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_and_loading_data.R
\name{skr_standard_uniform_names}
\alias{skr_standard_uniform_names}
\title{skr_standard_uniform_names}
\usage{
skr_standard_uniform_names(skr)
}
\description{
skr_standard_uniform_names
}
\keyword{internal}
