% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stat_per_obs_dataframe}
\alias{stat_per_obs_dataframe}
\title{Example dataset of stat_per_obs_dataframe}
\format{
\subsection{\code{stat_per_obs_dataframe}}{

A data.frame - it contains the provided factors, and the following
rows - each has 96 obs:
\describe{
\item{standardized_observedmean}{num  -0.847 -0.888 ...}
\item{standardized_min_quantilemean}{num  -1.95 -1.58 ...}
\item{standardized_max_quantilemean}{num  1.49 1.18 ...}
\item{significancemean}{logi  FALSE FALSE ...}
\item{standardized_observedvariance}{num  -0.635 -0.746 ...}
\item{standardized_min_quantilevariance}{num  -0.844 -1.183 ...}
\item{standardized_max_quantilevariance}{num  1.07 2.23 ...}
\item{significancevariance}{logi  FALSE FALSE ...}
\item{standardized_observedskewness}{num  0.904 1.458 ...}
\item{standardized_min_quantileskewness}{num  -1.21 -1.56 ...}
\item{standardized_max_quantileskewness}{num  1.23 1.43 ...}
\item{significanceskewness}{logi  FALSE TRUE ...}
\item{standardized_observedkurtosis}{num  1.61 3.17 ...}
\item{standardized_min_quantilekurtosis}{num  -0.654 -1.436 ...}
\item{standardized_max_quantilekurtosis}{num  0.996 1.717 ...}
\item{significancekurtosis}{logi  TRUE TRUE ...}
}
Example output of get_stat_per_rand
}
}
\usage{
stat_per_obs_dataframe
}
\description{
Example dataset of stat_per_obs_dataframe
}
\keyword{internal}
