% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stat_per_rand_dataframe}
\alias{stat_per_rand_dataframe}
\title{Example dataset of stat_per_rand_dataframe}
\format{
\subsection{\code{stat_per_rand_dataframe}}{

A data.frame - of 8 columns and 42 rows.
\describe{
\item{number}{int  0 0 1 1 2 2 3 3 4 4 ...}
\item{Treatment}{Factor w/ 2 levels}
\item{slope}{num  0.826 0.832 1.406 ...}
\item{intercept}{num  4.54 2.72 2.46 ...}
\item{rsquare}{num  0.201 0.216 ...}
\item{tad_stab}{num  1.856 0.49 ...}
\item{distance_to_family}{num  3.063 ...}
\item{cv_distance_to_family}{num  144 ...}
}
Example output of stat_per_rand_dataframe
}
}
\usage{
stat_per_rand_dataframe
}
\description{
Example dataset of stat_per_rand_dataframe
}
\keyword{internal}
