% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait}
\alias{trait}
\title{Example dataset containing some traits}
\format{
\subsection{\code{trait}}{

A data.frame with two columns and 98 rows:
\describe{
\item{Species}{chr sp1, sp2, ...}
\item{SLA}{num 15.4, 32, 32.5, ...}
}
}
}
\usage{
trait
}
\description{
Example dataset containing some traits
}
\keyword{internal}
