% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weighted_moments_dataframe}
\alias{weighted_moments_dataframe}
\title{Example dataset of weighted_moments_dataframe}
\format{
\subsection{\code{weighted_moments_dataframe}}{

A data.frame - todo = describe the df content:
Example output of weighted_moments_dataframe
}
}
\usage{
weighted_moments_dataframe
}
\description{
Example dataset of weighted_moments_dataframe
}
\keyword{internal}
