% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{weighted_mvsk}
\alias{weighted_mvsk}
\title{Compute the weighted mean, variance, skewness and kurtosis}
\usage{
weighted_mvsk(data, weights)
}
\arguments{
\item{data}{the data}

\item{weights}{the vector or matrix of weights corresponding to the
data (each row corresponding to an
iteration of data)}
}
\value{
the list of weighted mean, variance, skewness and
kurtosis of the data
}
\description{
Compute the weighted mean, variance, skewness and kurtosis
of data with given weights
}
\examples{

weighted_mvsk(
  data = c(1, 2, 3),
  weights = matrix(data = c(1, 1, 1, 2, 1, 3), nrow = 2, ncol = 3)
)

}
\concept{Statistics}
